Move URLBinding to its own file
