+++ adapter/internal/egress/syslog_connector.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"net/url"
 	"time"
 
 	"golang.org/x/net/context"
@@ -118,37 +117,6 @@ func WithLogClient(logClient LogClient) ConnectorOption {
 	return func(sc *SyslogConnector) {
 		sc.logClient = logClient
 	}
-}
-
-// URLBinding associates a particular application with a syslog URL. The
-// application is identified by AppID and Hostname. The syslog URL is
-// identified by URL.
-type URLBinding struct {
-	Context  context.Context
-	AppID    string
-	Hostname string
-	URL      *url.URL
-}
-
-// Scheme is a convenience wrapper around the *url.URL Scheme field
-func (u *URLBinding) Scheme() string {
-	return u.URL.Scheme
-}
-
-func buildBinding(b *v1.Binding, c context.Context) (*URLBinding, error) {
-	url, err := url.Parse(b.Drain)
-	if err != nil {
-		return nil, err
-	}
-
-	u := &URLBinding{
-		AppID:    b.AppId,
-		URL:      url,
-		Hostname: b.Hostname,
-		Context:  c,
-	}
-
-	return u, nil
 }
 
 // Connect returns an egress writer based on the scheme of the binding drain
