+++ adapter/internal/egress/url_binding.go
@@ -1,0 +1,39 @@
+package egress
+
+// URLBinding associates a particular application with a syslog URL. The
+import (
+	"context"
+	"net/url"
+
+	v1 "code.cloudfoundry.org/scalable-syslog/internal/api/v1"
+)
+
+// application is identified by AppID and Hostname. The syslog URL is
+// identified by URL.
+type URLBinding struct {
+	Context  context.Context
+	AppID    string
+	Hostname string
+	URL      *url.URL
+}
+
+// Scheme is a convenience wrapper around the *url.URL Scheme field
+func (u *URLBinding) Scheme() string {
+	return u.URL.Scheme
+}
+
+func buildBinding(b *v1.Binding, c context.Context) (*URLBinding, error) {
+	url, err := url.Parse(b.Drain)
+	if err != nil {
+		return nil, err
+	}
+
+	u := &URLBinding{
+		AppID:    b.AppId,
+		URL:      url,
+		Hostname: b.Hostname,
+		Context:  c,
+	}
+
+	return u, nil
+}
